{{ petor.project.slug }}/pyproject.toml
[project]
name = "{{ petor.project.name }}"
version = "0.1.0"
description = "{{ petor.project.description }}"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "collectfasta>=3.3.0",
  "django>=5.2.4",
  "django-environ>=0.12.0",
  "django-storages>=1.14.6",
  "djangorestframework>=3.16.0",
  "pillow>=11.3.0",
  "psycopg2>=2.9.10",
]
